requirements.txt
# MLOps Chatbot Project Requirements
# Python 3.10+

# Core ML/DL Frameworks
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.15.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# MLOps Tools
mlflow>=2.9.0
dvc>=3.30.0

# Serving & API
vllm>=0.2.6
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10

# UI/Demo
gradio>=4.7.0
streamlit>=1.28.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
pyyaml>=6.0.1
tenacity>=8.2.0

# Optional: For specific model downloads
huggingface-hub>=0.19.0
sentencepiece>=0.1.99
protobuf>=4.25.0
